+++ include/box2d/types.h
@@ -113,10 +113,13 @@ typedef struct b2WorldDef
 	float restitutionThreshold;
 
 	/// This parameter controls how fast overlap is resolved and has units of meters per second
-	float maxPushoutVelocity;
+	float contactPushoutVelocity;
 
-	/// This parameter affects the stiffness of contacts. Cycles per second.
+	/// Contact stiffness. Cycles per second.
 	float contactHertz;
+
+	/// Contact bounciness. Non-dimensional.
+	float contactDampingRatio;
 
 	/// Can bodies go to sleep to improve performance
 	bool enableSleep;
@@ -258,8 +261,9 @@ static inline b2WorldDef b2DefaultWorldDef(void)
 	b2WorldDef def = {0};
 	def.gravity = B2_LITERAL(b2Vec2){0.0f, -10.0f};
 	def.restitutionThreshold = 1.0f * b2_lengthUnitsPerMeter;
-	def.maxPushoutVelocity = 3.0f * b2_lengthUnitsPerMeter;
+	def.contactPushoutVelocity = 3.0f * b2_lengthUnitsPerMeter;
 	def.contactHertz = 30.0f;
+	def.contactDampingRatio = 1.0f;
 	def.enableSleep = true;
 	def.bodyCapacity = 8;
 	def.shapeCapacity = 8;
