+++ src/world.h
@@ -85,8 +85,9 @@ typedef struct b2World
 
 	b2Vec2 gravity;
 	float restitutionThreshold;
-	float maximumPushoutVelocity;
+	float contactPushoutVelocity;
 	float contactHertz;
+	float contactDampingRatio;
 
 	// This is used to compute the time step ratio to support a variable time step.
 	float inv_dt0;
